Finance/LogicRadar/requirements.txt
pandas
numpy
yfinance
fredapi
scikit-learn
openai
langchain
python-dotenv
streamlit
matplotlib
seaborn
requests
beautifulsoup4
lxml
openpyxl
pdfplumber
